worker_diarization/requirements_worker_diarization.txt
Flask==2.2.3
langchain==0.0.309
openai==0.27.7
pydub==0.25.1
werkzeug==2.2.3
flask-cors==4.0.0
google-cloud-storage
google-cloud-pubsub
google-cloud-firestore
google-cloud-tasks
deepgram-sdk
qdrant-client
tiktoken